getBySiteUserId 수정(테스트 완료)

+++ src/main/java/com/kakaotrack/choco/linkupapi/category/Category.java
@@ -14,7 +14,7 @@ import java.util.List;
 public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long category_id;
+    private int category_id;
 
     private String category_name;
